tft_dashboard/requirements.txt
psutil
pyserial
requests